% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyphreeqc-package.R
\docType{package}
\name{easyphreeqc}
\alias{easyphreeqc}
\alias{easyphreeqc-package}
\title{easyphreeqc}
\description{
This package is a high-level wrapper around the
\link[phreeqc]{phreeqc-package}, which provides a low-level interface to the
\href{https://wwwbrr.cr.usgs.gov/projects/GWC_coupled/phreeqc/}{phreeqc}
modelling environment.
}
\references{
Parkhurst DL and Apello CAJ. Description of Input and Examples
for PHREEQC. Version 3 - A Computer Program for Speciation, Batch-Reaction,
One-Dimensional Transport, and Inverse Geochemical Calculations. Available from
\url{https://wwwbrr.cr.usgs.gov/projects/GWC_coupled/phreeqc/phreeqc3-html/phreeqc3.htm}.
}
