% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc.R
\name{phreeqc}
\alias{phreeqc}
\title{Call phreeqc}
\usage{
phreeqc(..., db = NA)
}
\arguments{
\item{...}{A character vector(s) of inputs.}

\item{db}{A database to use (see \link{use_db}). Use NA for the default,
and NULL to skip (re)loading the database.}
}
\value{
A data.frame with the first \link{selected_output}
}
\description{
This function is a wrapper around \link[phreeqc]{phrRunString} and
\link[phreeqc]{phrGetSelectedOutput}. The database is reloaded by default,
because saved options lead to occasionally confusing output.
}
\examples{
phreeqc(phreeqc::ex2)

}
