% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc_input.R
\name{phreeqc_input}
\alias{phreeqc_input}
\alias{solution}
\alias{selected_output}
\alias{equilibrium_phases}
\alias{print.phreeqc_input}
\alias{is.phreeqc_input}
\alias{as.phreeqc_input}
\alias{+.phreeqc_input}
\title{Create input sections for phreeqc}
\usage{
phreeqc_input(type, number = NA, name = "", components = list())

solution(.number = 1, .name = "", ...)

selected_output(.number = 1, ...)

equilibrium_phases(.number = 1, .name = NA, ...)

\method{print}{phreeqc_input}(x, ...)

is.phreeqc_input(x)

as.phreeqc_input(x)

\method{+}{phreeqc_input}(x, y)
}
\arguments{
\item{type}{The keyword name to use (e.g. SOLUTION)}

\item{number, .number}{Number of the component}

\item{name, .name}{Name of the component}

\item{components, ...}{Key/value pairs corresponding to lines of input}

\item{x, y}{An object created by phreeqc_input}
}
\value{
A character vector with an element for each line
}
\description{
Create input sections for phreeqc
}
\examples{
# pure water solution
solution(pH=7, temp=25)

}
